requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
xgboost==2.0.3
scikit-learn==1.4.1.post1
pandas==2.2.1
numpy==1.26.4
joblib==1.3.2
python-dotenv==1.0.1
google-generativeai==0.5.2
httpx==0.27.0
